airflow/requirements.txt
beautifulsoup4
pandas
tqdm
numpy
selenium
# webdriver-manager
psycopg2-binary
requests
lxml

